refactor: Move login logic to client
