+++ cats/src/io/github/pidoveproject/showdown/cats/CatsShowdownConnection.scala
@@ -1,20 +1,16 @@
 package io.github.pidoveproject.showdown.cats
 
-import cats.effect.kernel.Concurrent
+import cats.effect.Concurrent
 import cats.syntax.all.*
 import fs2.Stream
 import io.github.iltotore.iron.*
-import io.github.pidoveproject.showdown.{ChallStr, ShowdownConnection}
-import io.github.pidoveproject.showdown.protocol.{LoginResponse, MessageInput, ProtocolError}
-import io.github.pidoveproject.showdown.protocol.client.{AuthCommand, ClientMessage}
+import io.github.pidoveproject.showdown.ShowdownConnection
+import io.github.pidoveproject.showdown.protocol.{MessageInput, ProtocolError}
+import io.github.pidoveproject.showdown.protocol.client.ClientMessage
 import io.github.pidoveproject.showdown.protocol.server.ServerMessage
 import io.github.pidoveproject.showdown.room.RoomId
-import io.github.pidoveproject.showdown.user.Username
-import org.http4s.{Method, Request, Uri, UrlForm}
-import org.http4s.EntityDecoder.text
 import org.http4s.client.Client
 import org.http4s.client.websocket.{WSConnection, WSFrame}
-import zio.json.*
 
 class CatsShowdownConnection[F[_]: Concurrent](client: Client[F], connection: WSConnection[F])
     extends ShowdownConnection[WSFrame, [e, r] =>> F[r], [r] =>> Stream[F, r]]:
@@ -60,30 +56,3 @@ class CatsShowdownConnection[F[_]: Concurrent](client: Client[F], connection: WS
               case Left(error) => Stream.emit(Left(ProtocolError.InvalidInput(text, error)))
 
         case _ => Stream.empty
-
-  override def login(challStr: ChallStr)(name: Username, password: String): F[LoginResponse] =
-    client.expect[String](
-      Request(
-        method = Method.POST,
-        uri = Uri.unsafeFromString("https://play.pokemonshowdown.com/action.php")
-      ).withEntity(UrlForm(
-        "act" -> "login",
-        "name" -> name.value,
-        "pass" -> password,
-        "challstr" -> challStr.value
-      )).asInstanceOf[Request[F]]
-    ).map(body => body.fromJson[LoginResponse].left.map(msg => ProtocolError.InvalidInput(body, msg)))
-      .rethrow
-      .flatTap(data => sendMessage(AuthCommand.Trn(name, 0, data.assertion)))
-
-  override def loginGuest(challStr: ChallStr)(name: Username): F[String] =
-    client.expect[String](
-      Request(
-        method = Method.POST,
-        uri = Uri.unsafeFromString("https://play.pokemonshowdown.com/action.php")
-      ).withEntity(UrlForm(
-        "act" -> "getassertion",
-        "userid" -> name.value,
-        "challstr" -> challStr.value
-      )).asInstanceOf[Request[F]]
-    ).flatTap(assertion => sendMessage(AuthCommand.Trn(name, 0, assertion)))
