+++ examples/zio-client/src/io/github/pidoveproject/showdown/client/package.scala
@@ -5,11 +5,11 @@ import zio.http.WebSocketFrame
 import zio.stream.Stream
 import io.github.pidoveproject.showdown.ShowdownConnection
 import io.github.pidoveproject.showdown.protocol.ProtocolError
-import io.github.pidoveproject.showdown.ZIOShowdownConnection
+import io.github.pidoveproject.showdown.{ZIOShowdownClient, ZIOShowdownConnection}
 
 /**
  * A protocol task depending on a [[ShowdownConnection]].
  * 
  * @tparam A the return type of this program
  */
-type ConnectionTask[+A] = ZIO[ZIOShowdownConnection, ProtocolError, A]
+type ConnectionTask[+A] = ZIO[ZIOShowdownClient & ZIOShowdownConnection, ProtocolError, A]
