+++ main/src/io/github/pidoveproject/showdown/protocol/CurrentUser.scala
@@ -1,6 +1,8 @@
 package io.github.pidoveproject.showdown.protocol
 
 import zio.json.*
+import io.github.iltotore.iron.zioJson.given
+import io.github.pidoveproject.showdown.user.Username
 
 @jsonMemberNames(CustomCase(_.toLowerCase))
-case class CurrentUser(loggedIn: Boolean, username: String, userId: String) derives JsonDecoder
+case class CurrentUser(loggedIn: Boolean, username: Username, userId: String) derives JsonDecoder
