+++ main/src/io/github/pidoveproject/showdown/protocol/LoginResponse.scala
@@ -1,9 +1,10 @@
 package io.github.pidoveproject.showdown.protocol
 
+import io.github.iltotore.iron.zioJson.given
 import zio.json.*
 
 case class LoginResponse(
     @jsonField("actionsuccess") actionSuccess: Boolean,
-    assertion: String,
+    assertion: Assertion,
     @jsonField("curuser") currentUser: CurrentUser
 ) derives JsonDecoder
